+++ src/app/components/admin/popups/userPopup/user-update-popup/user-update-popup.component.ts
@@ -11,6 +11,7 @@ import {MAT_DIALOG_DATA, MatDialog} from '@angular/material/dialog';
 import {User} from '../../../calendar/User';
 import {UserService} from '../../../../../service/UserService';
 import {MatCheckbox} from '@angular/material/checkbox';
+import {ConfirmationService} from '../../confirmation/confirmation.component';
 
 @Component({
   selector: 'app-user-update-popup',
@@ -36,8 +37,12 @@ import {MatCheckbox} from '@angular/material/checkbox';
 export class UserUpdatePopupComponent {
   updatedUser!: User;
   errorMessage:string = "";
-  constructor(private dialog: MatDialog, @Inject(MAT_DIALOG_DATA) protected user: User, private userService:UserService) {
-  }
+  constructor(
+    private dialog: MatDialog,
+    @Inject(MAT_DIALOG_DATA) protected user: User,
+    private userService:UserService,
+    private confirmationService: ConfirmationService,
+  ) {}
 
   ngOnInit() {
     this.updatedUser = {
@@ -57,40 +62,47 @@ export class UserUpdatePopupComponent {
     this.dialog.closeAll();
   }
   deleteUser(id:number) {
-    this.userService.deleteUserById(id).subscribe({
-        next:() =>{
-          window.location.reload()
-        },
-        error:(err) =>{
-          console.log(err)
-        }
+    this.confirmationService.performAction('Delete').subscribe(confirmation => {
+      if (confirmation) {
+        this.userService.deleteUserById(id).subscribe({
+            next:() =>{
+              window.location.reload()
+            },
+            error:(err) =>{
+              console.log(err)
+            }
+          }
+        );
       }
-
-    );
+    })
   }
   updateUser() {
-    const reservationRequest = {
-      id: this.updatedUser.id,
-      firstName: this.updatedUser.firstName,
-      lastName: this.updatedUser.lastName,
-      carRegistration: this.updatedUser.carRegistration,
-      city: this.updatedUser.city,
-      country: this.updatedUser.country,
-      email: this.updatedUser.email,
-      phoneNumber: this.updatedUser.phoneNumber,
-      reservations: this.updatedUser.reservations,
-      streetAddress: this.updatedUser.streetAddress
+    this.confirmationService.performAction('Update').subscribe(confirmation => {
+      if (confirmation) {
+        const reservationRequest = {
+          id: this.updatedUser.id,
+          firstName: this.updatedUser.firstName,
+          lastName: this.updatedUser.lastName,
+          carRegistration: this.updatedUser.carRegistration,
+          city: this.updatedUser.city,
+          country: this.updatedUser.country,
+          email: this.updatedUser.email,
+          phoneNumber: this.updatedUser.phoneNumber,
+          reservations: this.updatedUser.reservations,
+          streetAddress: this.updatedUser.streetAddress
+        }
+        this.userService.updateReservation(reservationRequest).subscribe({
+          next: () => {
+            console.log(reservationRequest)
+
+            window.location.reload();
+            this.closePopup();
+
+          },error:(err)=>{
+            this.errorMessage = err.error || 'Unexpected error...'
+          }
+        })
     }
-    this.userService.updateReservation(reservationRequest).subscribe({
-      next: () => {
-        console.log(reservationRequest)
-
-        window.location.reload();
-        this.closePopup();
-
-      },error:(err)=>{
-        this.errorMessage = err.error || 'Unexpected error...'
-      }
     })
   }
